Adds deployment sync lock
If a process tries to read from a netCDF file while it's being actively
written over, it will set a flag in the HDF5 library and prevent ALL
further read attempts on ANY files. Or at least that's what's happening
here.

This change includes a check for a file to help coordinate and
synchronize between the synchronization process and the application of
QARTOD