Adds attributes to QC variables
Sets _FillValue equal to -127 for any variables ending with `_qc` when
the attribute has not already been defined.

Automatically sets attributes for variables beginning with the string
`qartod`.  Variables automatically set are `flag_values`, `flag_meanings`,
and `_FillValue` (set to 2/NOT_EVALUATED, indicated QC has yet to be
performed)

Changes variables passed to template.  This is a breaking change and
requires the new templates.