Adds "compliance_check_passed" field, force option
Adds a new boolean field "compliance_check_passed" which indicates
whether or not the subset of checks run by the  Glider DAC compliance
checker passed or not.  By default any deployment that either does not
have the field present or failed the last check will be checked,
with passing deployments skipped over.  This can be overridden by
passing the `-f`/`--force` flag, which will check regardless of
the "compliance_check_passed" field.