Fixes/updates sorting of deployments
Fixes sorting of deployments corresponding to user and operator by turning
Mongo cursor results into a list, restoring legacy behavior.  Also
removes a potential error condition where if a deployment was in the
database but not in the corresponding location in the filesystem, trying
to sort by timestamp would have failed due to an attribute not existing.
New sorting code for listing deployments by user and operator now sorts
by timestamp first (falling back to 1970-01-01 as the implied timestamp
if not found), and deployment name second as the sorting keys.  Also changes
any deployment listing methods to use `list.sort()` instead of `sorted()` in
order to sort in place.