Refactor glider email code, add deployment check after submission
Refactors the glider email code to move the main logic into the
glider_dac/glider_emails.py file.  Also attempts to handle differing
aggregate syntax in MongoDB which is used to find the list of
deployments to run the metadata compliance check against.
Adds a glider deployment compliance check after submission and sends an
email to the deployment provider when finished.

Fixes compliance check scheduling inside "on_complete" method of
deployments.  The on_complete method does not only run on completion, but
also ran whenever the data was "synced".  This update checks only if the
deployment is actually marked completed **and** there wasn't an existing
check run as determined by the presence of the "compliance_check_passed"
field in the deployments collection.

For the compliance report, fetches the last day of ERDDAP data for each
deployment from the corresponding ERDDAP aggregation.  This greatly
reduces the needed amount of data to process while also improving the
checked metadata modifications from ERDDAP's datasets.xml.