Fix DB write contention for compliance check
Fixes an issue where there could be a race condition wherein a
queued compliance check which finished prior to a deployment save
operation would not write the result of the compliance check.  Fixed
by using update on individual fields since compliance checks do not need
to modify other fields of the deployment model.