Fixes issue with new deployment not finding id
Fixes an issue when a new deployment was created where a MongoDB id was
not assigned and threw a KeyError.  Adds exception handler where if "_id" field
isn't found, create a new deployment via "save" method.  This should be
fine with compliance checking functionality as a user first has to
create a deployment prior to adding files and marking as completed.