Pass compliance checking to scheduled jobs
Moves compliance checks on deployments to jobs scheduled half
an hour in the future.  If, on further invocations of checking
if the compliance check passed, if it is found that there is already
a job scheduled in the future for a particular deployment, the
compliance check is not scheduled again.  This cuts down on duplication
of jobs and the associated processing overhead due to many detected file
events while also allowing for conditions such as deployment data
resubmission to be handled.