Update glider_qc.py
Improved or added Docstring to function, method, or class

Used the logging module to log message at the error level

Added  a report to functions' returns to log in the error

Used fill_value=np.int16(-999) for _FillValue for qartod variables. Using a flag value to _FillValue is not a good practice for quality control data.

Add a full array of 2 as a default for qartod variables. There are instances where array cannot be qc tested because of errors in attributes values or there are insufficient data or the array is empty.

Add a functionality to drop configuration specifications and skip QC if the specs cannot be calculated. This is for the function get_spike_thresholds and  get_rate_of_change_threshold

Changed the qartod method to calculate flags from XarrayStream to PandasStream to fix the array type compatability issues with the ioos-qc module.

Added the values mask check and invalis data check to the time array.