modified apply_qc() to use PandasStore
Upgrading to ioos-qc version 2.2.0 caused a TypeError: 'ABCMeta' object is not subscriptable when running glider_qc_tests.py. The issue traced back to the aggregate method from the ioos-qc package. To resolve this, I modified glider_qc.py to use the PandasStore method for creating the aggregated flags.

Note: I have also added a TODO comments to set the extended file attribute as the current date-time.